+++ Anarchy/WebSockets/Gateway/Media/Voice/VoiceStateProperties.cs
@@ -47,5 +47,66 @@ namespace Discord.Gateway
             get { return VideoProperty; }
             set { VideoProperty.Value = value; }
         }
+
+        internal VoiceStateProperties Fill(DiscordSocketClient client)
+        {
+            if (client.Config.Cache)
+            {
+                DiscordVoiceStateContainer states;
+
+                try
+                {
+                    states = client.GetVoiceStates(client.User.Id);
+                }
+                catch (DiscordHttpException ex)
+                {
+                    if (ex.Code == DiscordError.UnknownUser)
+                        states = new DiscordVoiceStateContainer(client.User.Id);
+                    else
+                        throw;
+                }
+
+                DiscordVoiceState state = null;
+
+                if (GuildProperty.Set)
+                {
+                    if (GuildId.HasValue)
+                        states.GuildVoiceStates.TryGetValue(GuildId.Value, out state);
+                    else
+                        state = states.PrivateChannelVoiceState;
+
+                    if (state != null && !ChannelProperty.Set)
+                        ChannelId = state.Channel == null ? null : (ulong?)state.Channel.Id;
+                }
+                else if (ChannelProperty.Set && ChannelId.HasValue)
+                {
+                    var channel = client.GetChannel(ChannelId.Value);
+
+                    if (channel.Type != ChannelType.Group && channel.Type != ChannelType.DM)
+                    {
+                        GuildId = ((GuildChannel)channel).GuildId;
+
+                        if (states.GuildVoiceStates.TryGetValue(GuildId.Value, out DiscordVoiceState oldState))
+                            state = oldState;
+                    }
+                    else if (states.PrivateChannelVoiceState != null && states.PrivateChannelVoiceState.Channel != null && states.PrivateChannelVoiceState.Channel.Id == ChannelId)
+                        state = states.PrivateChannelVoiceState;
+                }
+
+                if (state != null)
+                {
+                    if (!DeafProperty.Set)
+                        Deafened = state.SelfDeafened;
+
+                    if (!MutedProperty.Set)
+                        Muted = state.Muted;
+
+                    if (!VideoProperty.Set)
+                        Video = state.Video;
+                }
+            }
+
+            return this;
+        }
     }
 }