+++ Anarchy/WebSockets/Media/Media/MediaSessionState.cs
@@ -2,8 +2,9 @@
 {
     public enum MediaSessionState
     {
-        NotConnected,
+        Dead,
+        StandBy,
         Connecting,
-        Connected
+        Authenticated
     }
 }